Camera and fade implemented
A character with a working camera has been implemented to be used on the game board overworld. I would need to set up the camera changing from player to player however. There is also a fade cube as an example object on how you can make other assets transparent if they get in the way of the camera