Instructions Screen(s) Ready up logic stuff
When the players start the game from the main menu, it will store how many players are playing the game.
This variable is used in the instructions screens and minigames (or eventually in minigames).
When all the players in the game ready up then the instructions screen will fade out and remove itself from the viewport - which will allow the minigame countdown to start and then begin.