Level Transitions
Added transitions for:
Lobby -> ArrowMiniGameLevel2
Lobby -> SpaceRace Minigame
Lobby -> Brawl (Vault room)

Ben is going to add the return after the minigames are complete. Programmers have already setup the game state for doing so, it just needs a finish condition and then to return now.

I've also added a Blueprint to Lobby's Level Blueprint as a quick idea/quality of life change. It makes it so that you no longer need to click on the window after hitting play for the controls to work properly/register.

(Do this by "Set Input Mode Game Only" after the Event BeginPlay, and connecting the Player Controller to Get Player Controller.)