level transitions + Camera
The camera in the vault and lobby have been adjusted sliughtly for better composition and focus.

You will now need to press E when overlapping a level transition trigger in the lobby as not to jar playtesters. It's set up to allow the A button on the gamepad to also do this, but is not curently working for unknown reasons.

You can now trigger the options hud with the "o" key.

Kelton approved camera changes for tomorrow's playtest.

Co-Authored-By: keltonlmorgan <keltonlmorgan@users.noreply.github.com>