File Cleanup + 3 States for Lobby (Read Description)
Removed majority of the Starter content. One or two things remain, as they are referenced in some of the BPs to do with the Blackhole, so I'll leave them for the time being.

Adjusted Lobby BPs to accomodate for the fact we're now having the Lobby as 3 separate levels.

Added some UI text that appears when you are within a trigger box that takes you to a Mini-game/other level -temporary for now, so feel free to have an art pass at that if needed/remove it later if no longer needed.

Ben & Kyle (or just one of them) now need to fix their references for the win states of their mini-games to now travel to the new states of the lobby instead. (Easy fix. Will take them two seconds.)

In general, just cleaned up the build a bit to lessen the folder size. Any issues; message me.