Adjusted text not overlapping
We're trying to get it so that only the first player is able to overlap with and activate the transitioning between levels.

The text UI is no longer overlapping multiple times on the same trigger box (for options). Will apply it to the rest of the trigger boxes.

The main issue is that we can't find the reference for the instance of the first player ID. Or at least, the method we're using isn't working properly.

Co-Authored-By: kylebottomley <kylebottomley@users.noreply.github.com>
Co-Authored-By: keltonlmorgan <keltonlmorgan@users.noreply.github.com>