Options menu functional
Able to navigate the options menu effectively as well as close it and resume gameplay.

A couple issues encountered: If player 1 is in range of the options desk (I assume any of the others too) any player can press Y to open the menu.

Second issue was that the volume mix override doesn't change the volume that the sounds play at.